update display functions
