+++ CMakeLists.txt
@@ -50,7 +50,6 @@ catkin_package(
 include_directories(include
   ${catkin_INCLUDE_DIRS}
   LibOVR/Include
-	src
   )
 
 IF(APPLE)
@@ -71,11 +70,11 @@ add_library(oculus_ros
   src/util.cpp
   )
 
-add_executable(oculus_ros_node src/oculus_ros_node.cpp)
-add_executable(distort_node src/distort.cpp src/distort_node.cpp src/viewer.cpp)
+add_executable(oculus_node src/oculus_node.cpp)
+add_executable(image_distort_viewer src/distort_image.cpp src/image_distort_viewer.cpp src/viewer.cpp)
 
 add_dependencies(oculus_ros oculus_ros_gencpp)
-add_dependencies(oculus_ros_node oculus_ros)
+add_dependencies(oculus_node oculus_ros)
 
 IF(APPLE)
   target_link_libraries(oculus_ros
@@ -84,8 +83,8 @@ ELSE(APPLE)
   target_link_libraries(oculus_ros udev Xinerama)
 ENDIF(APPLE)
 target_link_libraries(oculus_ros ovr ${catkin_LIBRARIES})
-target_link_libraries(distort_node ${catkin_LIBRARIES})
-target_link_libraries(oculus_ros_node oculus_ros)
+target_link_libraries(image_distort_viewer ${catkin_LIBRARIES})
+target_link_libraries(oculus_node oculus_ros)
 
 
 #############
@@ -103,7 +102,7 @@ target_link_libraries(oculus_ros_node oculus_ros)
 # )
 
 ## Mark executables and/or libraries for installation
-install(TARGETS oculus_ros oculus_ros_node
+install(TARGETS oculus_ros oculus_node image_distort_viewer
    ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
    LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
    RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
