+++ src/distort_image.cpp
