+++ src/image_distort_viewer.cpp
@@ -1,13 +1,13 @@
 #include <ros/ros.h>
 #include <oculus_ros/HMDInfo.h>
 #include <oculus_ros/distort.h>
-#include "viewer.h"
+#include <oculus_ros/viewer.h>
 
 namespace oculus_ros {
 
-class DistortNode {
+class ImageDistortViewer {
  public:
-  DistortNode();
+  ImageDistortViewer();
   void init();
   void HMDInfoCallback(const oculus_ros::HMDInfoPtr& info);
   void show();
@@ -15,20 +15,23 @@ class DistortNode {
   DistortImage left_;
   DistortImage right_;
   ros::Subscriber sub_;
+  bool use_display_;
   Viewer viewer_;
 };
 
-DistortNode::DistortNode()
-    : viewer_("oculus camera view") {
+ImageDistortViewer::ImageDistortViewer()
+    : use_display_(true)
+    , viewer_("oculus camera view")
+{
 }
 
-void DistortNode::init() {
+void ImageDistortViewer::init() {
 	ros::NodeHandle node;
   left_.init("camera_left/image_raw");
   right_.init("camera_right/image_raw");
   sub_ = node.subscribe("/oculus/hmd_info",
                         1,
-                        &oculus_ros::DistortNode::HMDInfoCallback,
+                        &ImageDistortViewer::HMDInfoCallback,
                         this);
   ros::NodeHandle private_node("~");
   int32_t offset_x = 0;
@@ -36,17 +39,21 @@ void DistortNode::init() {
   int32_t offset_y = 0;
   private_node.param<int32_t>("display_offset_y", offset_y, 0);
   viewer_.setDisplayOffset(offset_x, offset_y);
+  
+  private_node.param<bool>("use_display", use_display_, true);
 
 }
 
-void DistortNode::show() {
-  if ((!right_.getImage().empty()) &&
-      (!left_.getImage().empty())) {
-    viewer_.show(right_.getImage(), left_.getImage());
+void ImageDistortViewer::show() {
+  if (use_display_) {
+    if ((!right_.getImage().empty()) &&
+        (!left_.getImage().empty())) {
+      viewer_.show(right_.getImage(), left_.getImage());
+    }
   }
 }
 
-void DistortNode::HMDInfoCallback(const oculus_ros::HMDInfoPtr& info) {
+void ImageDistortViewer::HMDInfoCallback(const oculus_ros::HMDInfoPtr& info) {
   if (info->horizontal_screen_size > 0) {
     double lens_center = 1 - 2 * info->lens_separation_distance / info->horizontal_screen_size;
     double scale = 1 + lens_center;
@@ -65,7 +72,7 @@ void DistortNode::HMDInfoCallback(const oculus_ros::HMDInfoPtr& info) {
 int main(int argc, char** argv) {
 	ros::init(argc, argv, "oculus_distort_node");
   try {
-    oculus_ros::DistortNode dis;
+    oculus_ros::ImageDistortViewer dis;
     dis.init();
     while(ros::ok()) {
       cv::waitKey(10);
